Moving IValidationStateService and impl to Validation.PackageSigning.Core project.
