+++ src/Validation.PackageSigning.Core/Storage/IValidatorStateService.cs
@@ -2,44 +2,10 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System.Threading.Tasks;
+using NuGet.Services.Validation;
 
-namespace NuGet.Services.Validation.Orchestrator
+namespace NuGet.Jobs.Validation.PackageSigning.Storage
 {
-    /// <summary>
-    /// The possible results for <see cref="IValidatorStateService.AddStatusAsync(string, ValidatorStatus)"/>.
-    /// </summary>
-    public enum AddStatusResult
-    {
-        /// <summary>
-        /// Successfully persisted the <see cref="ValidatorStatus"/>.
-        /// </summary>
-        Success,
-
-        /// <summary>
-        /// Failed to persist the <see cref="ValidatorStatus"/> as a status already
-        /// exists with the same validation id.
-        /// </summary>
-        StatusAlreadyExists,
-    }
-
-    /// <summary>
-    /// The possible results for <see cref="IValidatorStateService.SaveStatusAsync(string, ValidatorStatus)"/>.
-    /// </summary>
-    public enum SaveStatusResult
-    {
-        /// <summary>
-        /// Successfully persisted the updated <see cref="ValidatorStatus"/>
-        /// </summary>
-        Success,
-
-        /// <summary>
-        /// The <see cref="ValidatorStatus"/> is stale. The status should be refetched using
-        /// <see cref="IValidatorStateService.GetStatusAsync(string, IValidationRequest)"/> before attempting
-        /// to save again.
-        /// </summary>
-        StaleStatus,
-    }
-
     /// <summary>
     /// A service used to persist a <see cref="IValidator"/>'s validation statuses.
     /// </summary>
