+++ src/Validation.PackageSigning.Core/Validation.PackageSigning.Core.csproj
@@ -33,21 +33,24 @@
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.Core" />
-    <Reference Include="System.Xml.Linq" />
-    <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
     <Reference Include="System.Data" />
-    <Reference Include="System.Net.Http" />
-    <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="NuGet.Services.ServiceBus" Version="2.3.2-master-15927" />
+    <PackageReference Include="NuGet.Services.Validation">
+      <Version>2.3.2-master-15927</Version>
+    </PackageReference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ICertificateStore.cs" />
+    <Compile Include="Storage\AddStatusResult.cs" />
+    <Compile Include="Storage\ICertificateStore.cs" />
+    <Compile Include="Storage\IValidatorStateService.cs" />
     <Compile Include="Messages\CertificateValidationMessage.cs" />
     <Compile Include="Messages\CertificateValidationMessageSerializer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Storage\SaveStatusResult.cs" />
+    <Compile Include="Storage\ValidatorStateService.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
