+++ tests/NuGet.Services.Validation.Orchestrator.Tests/PackageSigning/PackageSigningValidatorFacts.cs
@@ -7,6 +7,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Moq;
+using NuGet.Jobs.Validation.PackageSigning.Storage;
 using NuGet.Services.Validation.Orchestrator;
 using Xunit;
